+++ example/lib/view/widget/ChatDetailWidget.dart
@@ -1,5 +1,5 @@
-import 'package:fllama_example/view/widget/ChatListItemWidget.dart';
-import 'package:fllama_example/vm/HomeVM.dart';
+import 'package:fcllama_example/view/widget/ChatListItemWidget.dart';
+import 'package:fcllama_example/vm/HomeVM.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';