+++ example/lib/view/widget/ChatInputWidget.dart
@@ -1,4 +1,4 @@
-import 'package:fllama_example/vm/HomeVM.dart';
+import 'package:fcllama_example/vm/HomeVM.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:get/get.dart';