+++ example/lib/vm/AboutVM.dart
@@ -1,4 +1,4 @@
-import 'package:fllama/fllama.dart';
+import 'package:fcllama/fllama.dart';
 import 'package:get/get.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
@@ -14,7 +14,7 @@ class AboutVM extends GetxController {
           buildNumber.value = value.buildNumber.toString(),
           Get.log("V${version.value}[${buildNumber.value}]")
         });
-    Fllama.instance()?.getCpuInfo().then((value) {
+    FCllama.instance()?.getCpuInfo().then((value) {
       Get.log("CPU-Info=$value");
       cpuInfo.value = value.toString();
     });