+++ CHANGELOG.md
@@ -1,8 +1,28 @@
 ## 0.0.1
 
-The first official Flutter package of HLlama.
+The first official Flutter package of Llama.
 - llama.cpp based on [54ef9cfc726a799e6f454ac22c4815d037716eda](https://github.com/ggerganov/llama.cpp/commits/master/)
 * Optimizing performance
 * AGP8+ compatible
-* See the example project for more details.
+* See the example project for more details
 - 2024-11-12
+
+## 0.0.2
+
+Package rename to FCLlama , the first official flutter package of Llama , which use platform channel.
+- llama.cpp based on [6dfcfef0787e9902df29f510b63621f60a09a50b](https://github.com/ggerganov/llama.cpp/commits/master/)
+* Optimizing performance
+* AGP8+ compatible
+* See the example project for more details
+* Update NDK version and llama.cpp version
+- 2024-11-23
+
+## 0.0.3
+
+- llama.cpp based on [6dfcfef0787e9902df29f510b63621f60a09a50b](https://github.com/ggerganov/llama.cpp/commits/master/)
+* Optimizing performance
+* AGP8+ compatible
+* See the example project for more details
+* Update NDK version and llama.cpp version
+* Update README.md
+- 2024-11-23